feat/fix: set the theme to One Dark to mitigate highlighting issues
It is one of the most popular and well-looking custom themes out there,
which takes its roots in the Atom editor (when it was active).

It's doing much better job at using different colors for different
syntax constructs, so I decided to go with it over hacking together
custom CSS theme for our docs.